requirements.txt
mysql-connector-python
bcrypt
opencv-python
numpy
cmake
dlib
face-recognition
SpeechRecognition
reportlab
pyaudio
pyttsx3
pqueue